Rename `test` folder to `tests` and autoload it
