+++ tests/MetricsTest.php
@@ -1,12 +1,16 @@
 <?php
 
-class MetricsTest extends Delighted\TestCase
+namespace Delighted\Tests;
+
+use Delighted\Metrics;
+
+class MetricsTest extends TestCase
 {
 
     public function testRetrievingMetrics()
     {
         $this->addMockResponse(200, json_encode(['nps' => 10]));
-        $metrics = Delighted\Metrics::retrieve();
+        $metrics = Metrics::retrieve();
         $this->assertInstanceOf('Delighted\Metrics', $metrics);
         $this->assertObjectPropertyIs(10, $metrics, 'nps');
         $this->assertObjectNotHasAttribute('id', $metrics);
