+++ tests/TestClient.php
@@ -1,11 +1,13 @@
 <?php
 
-namespace Delighted;
+namespace Delighted\Tests;
+
+use Delighted\Client;
 
 class TestClient extends Client
 {
     public function getAdapter()
     {
-        return $this->adapter;
+        return $this->httpClient;
     }
 }
